Temporally replaces `*` for `+` regexp for matching digits.
With an asterisk, the regexp matches a single `-` as a digit. That
asterisk is used to match _zero_ or more digits before a decimal point,
so a value like `-.4` is accepted as number.

But due the fact that the decimal point and all the decimal digits
should be optional, it accepts a single `-` as a digit, causing this
bug.

This commit replaces the `*` for a `+` so always an integer part is
required, thus invalidating this (frankly